working on the string

+++ phpFiles/api.php
@@ -16,7 +16,7 @@
 		
 		//query
 		$getRoleQuery = "SELECT USER_ROLE FROM User WHERE USER_ID = '$usersID' AND USER_EMAIL = '$usersEmail'";
-		$deleteUserQuery = "DELETE FROM User WHERE USER_ID = '$usersID' AND USER_EMAIL = '$usersEmail';";
+		$deleteUserQuery = "DELETE FROM User WHERE User.USER_ID = '$usersID' AND User.USER_EMAIL = '$usersEmail';";
 		$deleteQuery = "";
 		$didItWork = "";
 		
@@ -28,21 +28,21 @@
 		$array = mysqli_fetch_row($usersRoleResult);
 		
 		if($array[0] == 1){
-			$deleteQuery = "DELETE FROM Administrator WHERE USER_ID = '$usersID'; " . $deleteUserQuery;
+			$deleteQuery = "DELETE FROM Administrator WHERE Administrator.USER_ID = '$usersID'; " . $deleteUserQuery;
 			if (!mysqli_query($con,$deleteQuery)){
 				$didItWork = mysqli_error($con);
   			}		
 		}
 		
 		if($array[0] == 2){
-			$deleteQuery = "DELETE FROM Teacher WHERE USER_ID = '$usersID'; " . $deleteUserQuery;
+			$deleteQuery = "DELETE FROM Teacher WHERE Teacher.USER_ID = '$usersID'; " . $deleteUserQuery;
 			if (!mysqli_query($con,$deleteQuery)){
 				$didItWork = mysqli_error($con);
   			}		
   		}
   		
 		if($array[0] == 3){
-			$deleteQuery = "DELETE FROM Student WHERE USER_ID = '$usersID'; " . $deleteUserQuery;
+			$deleteQuery = "DELETE FROM Student WHERE Student.USER_ID = '$usersID'; " . $deleteUserQuery;
 			if (!mysqli_query($con,$deleteQuery)){
 				$didItWork = mysqli_error($con);
   			}
